Extract backup functions into a dedicated service
