+++ zou/app/services/sync_service.py
@@ -457,28 +457,6 @@ def add_sync_listeners(event_client, model_name, event_name, model):
     )
 
 
-def add_file_listeners(event_client):
-    """
-    Add new preview event listener.
-    """
-    gazu.events.add_listener(
-        event_client, "preview-file:add-file", add_file_event
-    )
-
-
-def add_file_event(data):
-    if data.get("sync", False):
-        return
-    try:
-        preview_file_id = data["preview_file_id"]
-        preview_file = PreviewFile.get(preview_file_id)
-        download_preview_from_another_instance(preview_file)
-        logger.info("Preview file and related downloaded: %s" % preview_file_id)
-    except gazu.exception.RouteNotFoundException as e:
-        logger.error("Route not found: %s" % e)
-        logger.error("Fail to dowonload preview file: %s" % (preview_file_id))
-
-
 def create_entry(model_name, event_name, model, event_type):
     """
     Generate a function that creates a model each time a related creation event
@@ -549,6 +527,29 @@ def forward_base_event(event_name, event_type, data):
     data["sync"] = True
     logger.info("Forward event: %s" % full_event_name)
     events.emit(full_event_name, data)
+
+
+def add_file_listeners(event_client):
+    """
+    Add new preview event listener.
+    """
+    gazu.events.add_listener(
+        event_client, "preview-file:add-file", add_file_event
+    )
+
+
+def add_file_event(data):
+    if data.get("sync", False):
+        return
+    try:
+        preview_file_id = data["preview_file_id"]
+        preview_file = PreviewFile.get(preview_file_id)
+        download_preview_from_another_instance(preview_file)
+        logger.info("Preview file and related downloaded: %s" % preview_file_id)
+    except gazu.exception.RouteNotFoundException as e:
+        logger.error("Route not found: %s" % e)
+        logger.error("Fail to dowonload preview file: %s" % (preview_file_id))
+
 
 
 def download_entity_thumbnails_from_storage():
@@ -643,139 +644,10 @@ def download_preview(preview_file):
     download_file(file_path, "previews", dl_func, preview_file_id)
 
 
-def generate_db_backup(host, port, user, password, database):
-    """
-    Generate a Postgres dump file from the database.
-    """
-    now = datetime.datetime.now().strftime("%Y-%m-%d")
-    filename = "%s-zou-db-backup.dump" % now
-    with gzip.open(filename, "wb") as archive:
-        pg_dump(
-            "-h",
-            host,
-            "-p",
-            port,
-            "-U",
-            user,
-            database,
-            _out=archive,
-            _env={"PGPASSWORD": password},
-        )
-    return filename
-
-
-def store_db_backup(filename):
-    """
-    Store given file located in the same directory, inside the files bucket
-    using the `dbbackup` prefix.
-    """
-    from zou.app import app
-
-    with app.app_context():
-        file_store.add_file("dbbackup", filename, filename)
-
-
-def upload_entity_thumbnails_to_storage(days=None):
-    """
-    Upload all thumbnail files for non preview entries to object storage.
-    """
-    upload_entity_thumbnails(Project, days)
-    upload_entity_thumbnails(Organisation, days)
-    upload_entity_thumbnails(Person, days)
-
-
-def upload_entity_thumbnails(model, days=None):
-    query = model.query
-    if days is not None:
-        limit_date = date_helpers.get_date_from_now(int(days))
-        query = query.filter(model.updated_at >= limit_date)
-
-    for entity in query.all():
-        upload_entity_thumbnail(entity)
-
-
-def upload_preview_files_to_storage(days=None):
-    """
-    Upload all thumbnail and original files for preview entries to object
-    storage.
-    """
-    query = PreviewFile.query
-    if days is not None:
-        limit_date = date_helpers.get_date_from_now(int(days))
-        query = query.filter(PreviewFile.updated_at >= limit_date)
-
-    for preview_file in query.all():
-        upload_preview(preview_file)
-
-
-def upload_entity_thumbnail(entity):
-    """
-    Upload thumbnail file for given entity to object storage.
-    """
-    preview_folder = os.getenv("PREVIEW_FOLDER", "/opt/zou/previews")
-    local = LocalBackend(
-        "local", {"root": os.path.join(preview_folder, "pictures")}
-    )
-
-    file_path = local.path("thumbnails-" + str(entity.id))
-    if entity.has_avatar:
-        file_store.add_picture("thumbnails", str(entity.id), file_path)
-        print("%s uploaded" % file_path)
-
-
-def upload_preview(preview_file):
-    """
-    Upload all files link to preview file entry: orginal file and variants.
-    """
-    print("upload preview %s (%s)" % (preview_file.id, preview_file.extension))
-
-    preview_folder = os.getenv("PREVIEW_FOLDER", "/opt/zou/previews")
-    local_picture = LocalBackend(
-        "local", {"root": os.path.join(preview_folder, "pictures")}
-    )
-    local_movie = LocalBackend(
-        "local", {"root": os.path.join(preview_folder, "movies")}
-    )
-    local_file = LocalBackend(
-        "local", {"root": os.path.join(preview_folder, "files")}
-    )
-
-    is_movie = preview_file.extension == "mp4"
-    is_picture = preview_file.extension == "png"
-    is_file = not is_movie and not is_picture
-
-    preview_file_id = str(preview_file.id)
-    file_key = "previews-%s" % preview_file_id
-    if is_picture:
-        file_path = local_picture.path(file_key)
-        ul_func = file_store.add_picture
-        exists_func = file_store.exists_picture
-    elif is_movie:
-        file_path = local_movie.path(file_key)
-        ul_func = file_store.add_movie
-        exists_func = file_store.exists_movie
-    elif is_file:
-        file_path = local_file.path(file_key)
-        ul_func = file_store.add_file
-        exists_func = file_store.exists_file
-
-    if is_movie or is_picture:
-        for prefix in ["thumbnails", "thumbnails-square", "original"]:
-            pic_file_path = local_picture.path(
-                "%s-%s" % (prefix, str(preview_file.id))
-            )
-            if os.path.exists(pic_file_path) and not file_store.exists_picture(
-                prefix, preview_file_id
-            ):
-                file_store.add_picture(prefix, preview_file_id, pic_file_path)
-
-    prefix = "previews"
-    if os.path.exists(file_path) and not exists_func(prefix, preview_file_id):
-        ul_func(prefix, preview_file_id, file_path)
-    print("%s uploaded" % file_path)
-
-
 def download_files_from_another_instance():
+    """
+    Download all files from target instance.
+    """
     download_thumbnails_from_another_instance("person")
     download_thumbnails_from_another_instance("organisation")
     download_thumbnails_from_another_instance("project")
@@ -784,6 +656,9 @@ def download_files_from_another_instance():
 
 
 def download_thumbnails_from_another_instance(model_name):
+    """
+    Download all thumbnails from target instance for given model.
+    """
     model = event_name_model_map[model_name]
     for instance in model.query.all():
         if instance.has_avatar:
@@ -791,6 +666,9 @@ def download_thumbnails_from_another_instance(model_name):
 
 
 def download_thumbnail_from_another_instance(model_name, model_id):
+    """
+    Download into the local storage the thumbnail for a given model instance.
+    """
     file_path = local_picture.path("thumbnails-%s" % str(model_id))
     dirname = os.path.dirname(file_path)
     if not os.path.exists(dirname):
